Refactor SQL conversion routines into a separate file for better organization
