--- lib/production_sampler/preload_hash.rb
@@ -1,1 +1,0 @@
-# TODO refactor all preloading routines into this file
